// src/Client/TreeMon.Mobile/src/app/models/serviceresult.ts
// Copyright 2015, 2017 TreeMon.org.
// Licensed under CPAL 1.0,  See license.txt  or go to http://treemon.org/docs/license.txt  for full license details.

export class ServiceResult {

    RoleWeight: number;

    Code: number;

    Message: string;

    Result: any;

    Status: string;

    TotalRecordCount: number;

    // for image upload
    ImageUrl: string;

    ImageThumb: string;
}